Adds donation impact section
